requirements.txt
py-clob-client
python-dotenv
websockets>=12.0
